# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for Signal Bot",
        "description": "API documentation",
        "contact": {
            "email": "kenwaribo@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:8000/",
            "description": "Returns App API"
        }
    ],
    "paths": {
        "/": {},
        "/version": {
            "get": {
                "summary": "View installed version of Signal CLI",
                "operationId": "f19245d4fa1003ff21b975ff983962f0",
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    }
                }
            }
        },
        "/receive": {
            "get": {
                "summary": "Receieve unread messages",
                "operationId": "25550291f427d9974d90ca86dd250dce",
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    }
                }
            }
        },
        "/verify": {
            "post": {
                "summary": "Verify a registered phone number using otp from sms",
                "description": "Send params in form-data. Add Accept:application/json in header",
                "operationId": "verify",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "otp code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/sendmessage": {
            "post": {
                "summary": "Send a message to user",
                "description": "Send params in form-data. Add Accept:application/json in header",
                "operationId": "sendmessage",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "receipient phone number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "message",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/profilename": {
            "post": {
                "summary": "Update profile name of registered phone number",
                "description": "Send params in form-data. Add Accept:application/json in header",
                "operationId": "profilename",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name string",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Endpoint to register phone number. Number has to be set in .env because it's better to use one number at a time with signal",
                "description": "Send params in form-data. Add Accept:application/json in header",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "captcha",
                        "in": "query",
                        "description": "captcha gottent from https://signalcaptchas.org/registration/generate.html or https://signalcaptchas.org/challenge/generate.html",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "captcha",
                        "in": "query",
                        "description": "captcha",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}